Device Рефакторинг

+++ sources/Device/src/Multimeter/Multimeter.cpp
@@ -96,7 +96,7 @@ void Multimeter::Update()
     }
     
     uint8 range = 0;
-    if(MULTI_MEASURE == Measure::VoltageDC)        { range = MULTI_RANGE_DC; }
+    if(MULTI_MEASURE == Measure::VoltageDC)        { range = (uint8)MULTI_RANGE_DC; }
     else if(MULTI_MEASURE == Measure::VoltageAC)   { range = (uint8)MULTI_RANGE_AC; }
     else if(MULTI_MEASURE == Measure::Resistance)  { range = (uint8)MULTI_RANGE_RESISTANCE; }
     else
